Clean up KS2 (we like the new implementation!)
