+++ spikesorters/kilosort2/__init__.py
@@ -1,2 +1,1 @@
 from .kilosort2 import Kilosort2Sorter
-from .kilosort2_new import Kilosort2SorterNew
